requirements/requirements.txt
aiohttp==3.11.9
aiolimiter==1.2.1
anthropic==0.43.1
bitsandbytes==0.45.0
decord==0.6.0
deepspeed==0.16.2
einops==0.8.0
fastapi==0.115.6
gradio==5.12.0
huggingface_hub #==0.24.6
imageio==2.37.0
llava==0.0.1.dev0
numpy==2.2.2
openai==1.59.9
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
peft==0.14.0
Pillow==11.1.0
# protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.5
pysubs2==1.8.0
python-dotenv==1.0.1
Requests==2.32.3
sglang==0.4.1.post7
spaces==0.32.0
tabulate==0.9.0
timm==1.0.14
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.51.1 #4.48.0
uvicorn==0.34.0
vllm==v0.8.5.post1 #0.8.2 #==0.6.6.post2.dev270+g630eb5b5
# google-generativeai
google-genai